move ColumnIndexAttribute to primitives, nix path attr
