+++ crates/bindings-macro/Cargo.toml
@@ -11,6 +11,8 @@ proc-macro = true
 bench = false
 
 [dependencies]
+spacetimedb-primitives = { path = "../primitives", version = "0.7.3" }
+
 bitflags.workspace = true
 humantime.workspace = true
 proc-macro2.workspace = true
