+++ crates/bindings/src/lib.rs
@@ -10,34 +10,31 @@ pub mod rt;
 pub mod time_span;
 mod timestamp;
 
+use crate::sats::db::def::IndexType;
 use spacetimedb_lib::buffer::{BufReader, BufWriter, Cursor, DecodeError};
-pub use spacetimedb_lib::de::{Deserialize, DeserializeOwned};
 use spacetimedb_lib::sats::{impl_deserialize, impl_serialize, impl_st};
-pub use spacetimedb_lib::ser::Serialize;
 use spacetimedb_lib::{bsatn, PrimaryKey, ProductType, ProductValue};
+use spacetimedb_primitives::ColumnIndexAttribute;
 use std::cell::RefCell;
 use std::marker::PhantomData;
 use std::slice::from_ref;
 use std::{fmt, panic};
+use sys::{Buffer, BufferIter};
 
-pub use spacetimedb_bindings_macro::{duration, query, spacetimedb, TableType};
-
+pub use log;
 pub use sats::SpacetimeType;
+pub use spacetimedb_bindings_macro::{duration, query, spacetimedb, TableType};
+pub use spacetimedb_bindings_sys as sys;
 pub use spacetimedb_lib;
+pub use spacetimedb_lib::de::{Deserialize, DeserializeOwned};
 pub use spacetimedb_lib::sats;
+pub use spacetimedb_lib::ser::Serialize;
 pub use spacetimedb_lib::Address;
 pub use spacetimedb_lib::AlgebraicValue;
 pub use spacetimedb_lib::Identity;
 pub use spacetimedb_primitives::TableId;
-pub use timestamp::Timestamp;
-
-pub use spacetimedb_bindings_sys as sys;
 pub use sys::Errno;
-use sys::{Buffer, BufferIter};
-
-use crate::sats::db::attr::ColumnIndexAttribute;
-use crate::sats::db::def::IndexType;
-pub use log;
+pub use timestamp::Timestamp;
 
 pub type Result<T = (), E = Errno> = core::result::Result<T, E>;
