+++ crates/core/src/db/relational_db.rs
@@ -15,9 +15,8 @@ use crate::execution_context::ExecutionContext;
 use fs2::FileExt;
 use nonempty::NonEmpty;
 use spacetimedb_lib::PrimaryKey;
-use spacetimedb_primitives::{ColId, IndexId, SequenceId, TableId};
+use spacetimedb_primitives::{ColId, ColumnIndexAttribute, IndexId, SequenceId, TableId};
 use spacetimedb_sats::data_key::ToDataKey;
-use spacetimedb_sats::db::attr::ColumnIndexAttribute;
 use spacetimedb_sats::db::def::{IndexDef, SequenceDef, TableDef, TableSchema};
 use spacetimedb_sats::hash::Hash;
 use spacetimedb_sats::{AlgebraicType, AlgebraicValue, ProductType, ProductValue};
