+++ crates/primitives/src/lib.rs
@@ -1,5 +1,7 @@
 #![no_std]
 
 mod ids;
+mod attr;
 
 pub use ids::{ColId, ConstraintId, IndexId, SequenceId, TableId};
+pub use attr::{AttributeKind, ColumnIndexAttribute};
